Analyze CSV file to find out and preserve dialect
The CSV export for an account statement may have a different delimiter
than the default (',') or have specific quoting rules.

When re-creating the CSV for the CSV tag in the ledger transaction the
original format should be preserved in order to be able to recreate
the CSV from the ledger file.

This will help when verifying the generated ledger transaction with the
original CSV source using the MD5Sum tag.

Add output_tag_csv_quoting option to allow specific quoting rules for
the generated CSV tag, to further match the original CSV more appropriately
if it contains custom quoting rules (e.g. QUOTE_ALL).