Replace date_format with csv_date_format and ledger_date_format
By default the same date format from the CSV is be used.

If the user would like the ledger file to use a different format,
the date is parsed from the CSV using csv_date_format and printed
to the ledger file using ledger_date_format.

This allows the user to for further customize of the generated
ledger file.