Determine CSV dialect and preserve original CSV
The CSV export for an account statement may have a different delimiter
than the default (','), icsv2ledger analyzes the CSV to determine the
specific dialect used for parsing.  This adds support for various CSV formats.

The value of the CSV tag is now the original CSV line as it was read,
this will help when verifying the generated ledger transaction with
the original CSV source using the MD5Sum tag.