Update to python3 syntax
Assume all input and output, other than the input CSV file, is encoded
in UTF-8. The ledger-cli documentation specifies that all ledger files
are encoded with UTF-8. It only makes sense for all other icsv2ledger
files (templates, mappings) to be UTF-8 as well. The encoding for the
input CSV file is left unspecified so that it will default to the user's
locale encoding.

**NOTE**: If the user has been using icsv2ledger in a non-UTF-8 locale
with non-ASCII CSV files then they will have to manually convert their
template and mapping files (and possibly ledger files, if they haven't
been keeping them in UTF-8 as they should) to UTF-8 encoding before
updating to this version of icsv2ledger.