make formatting of all tags on a single line optional
Tags inlining is not always welcome. In particular, inlining metadata tags
break their semantics, as the value of a metadata tag extends until the end of
line.

With this change a new command line option --multiline-tags and a corresponding
configuration file option are added to alter the inlining behavior. The default
remains unchanged to inlining tags by default.

Closes #103